broke hamburger button test out to seperate file
